backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.4.0
aiofiles==23.2.1
python-dotenv==1.0.0
